requirements.txt
# ==============================
# Core Data Processing
# ==============================
pandas==2.0.3
numpy==1.24.3
pyarrow==13.0.0

# ==============================
# Apache Kafka
# ==============================
kafka-python==2.0.2
confluent-kafka==2.3.0

# ==============================
# Apache Spark
# ==============================
pyspark==3.5.0

# ==============================
# Elasticsearch
# ==============================
elasticsearch==8.11.0

# ==============================
# Apache Cassandra
# ==============================
cassandra-driver==3.28.0

# ==============================
# Machine Learning
# ==============================
scikit-learn==1.3.2
mlflow==2.9.0
joblib==1.3.2

# ==============================
# Data Generation & Utilities
# ==============================
faker==20.1.0
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# ==============================
# HTTP & API
# ==============================
requests==2.31.0
urllib3==2.1.0

# ==============================
# Testing
# ==============================
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# ==============================
# Code Quality
# ==============================
black==23.12.0
flake8==6.1.0
pylint==3.0.3
isort==5.13.2

# ==============================
# Development Tools
# ==============================
ipython==8.18.1
jupyter==1.0.0
notebook==7.0.6

# ==============================
# Monitoring & Logging
# ==============================
prometheus-client==0.19.0
python-json-logger==2.0.7

# ==============================
# Date & Time
# ==============================
python-dateutil==2.8.2
pytz==2023.3.post1

# ==============================
# Serialization
# ==============================
avro-python3==1.10.2
msgpack==1.0.7

# ==============================
# Type Hints
# ==============================
typing-extensions==4.9.0
